CogVM source as per VMMaker.oscog-eem.1687
Spur:
Fix bug in markAndTraceWeaklingsFrom:.  Fields of weaklings may be immediate
so must use followOopField:ofObject: not followObjField:ofObject:.

Slang:
Emit constants before types to allow constant names to be used when defining
types, e.g. for array extents.

Threaded FFI plugin:
Use above facility for neater declarations.

First cut (untested) of FFI support for x86-64 on System V and WIN64, as
X64SysVFFIPlugin.c & X64Win64FFIPlugin.c.

Surface Plugin:
Change int to long throughout parameter passing (but not the
Surface struct itself) to solve issues on 64-bits.

Mac OS:
remember localizable strings versions for Pharo and Newspeak.
Change some (int) casts to (long) or non casts in OpenGL and event code, again
to avoid issues on 64-bits.


git-svn-id: http://squeakvm.org/svn/squeak/branches/Cog@3617 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: e6bcbc14230fe81297864e1cdd0184054a2a8ef1

+++ platforms/iOS/vm/OSX/sqSqueakOSXApplication+events.m
@@ -392,7 +392,7 @@ static int buttonState=0;
 	}
 
 	// button state: low three bits are mouse buttons; next 8 bits are modifier bits
-	buttonState = (int) (modifierMap[((keyBoardCarbonModifiers & 0xFFFF) >> 8)] << 3) | (int) (stButtons & 0x7);
+	buttonState = (modifierMap[((keyBoardCarbonModifiers & 0xFFFF) >> 8)] << 3) | (stButtons & 0x7);
 	return buttonState;
 }
